+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- React `ref` forwarding for `<Field />`, `<Form />` and `<Message />`
+
 ## [0.0.8] - 2025-04-24
 
 ### Added